terrenos_panoramicos/requirements.txt
asgiref==3.3.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.1
django-storages==1.13.2
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-core==2.3.3
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==8.2.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-decouple==3.8
pytz==2021.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.16
django-jsonify